Update LV bus + line comparison data for test_run_powerflow()
With changing how generators and loads in buildings are connected to a
bus (now only one common building bus instead of a separate one for each
load and each generator in the PyPSA export) the resulting data changes.

This affects the total number of lines and buses in the comparison data
as well as the voltage and power values.